refactor: generify `SortCallback` interface
