+++ app/src/main/java/com/hxl/coinfuse/ui/fragments/navigation/NavigationFragment.java
@@ -21,7 +21,7 @@ public class NavigationFragment extends Fragment {
     public static final String exchangeArgKey = "exchangeId";
     public static final String explorerArgKey = "explorerId";
     public static final String isTimeSortableArgKey = "isTimeSortable";
-    public static final String coinSortCallbackArgKey = "coinSortCallback";
+    public static final String sortCallbackArgKey = "sortCallback";
     public static final String orderByArgKey = "orderBy";
     public static final String sortByArgKey = "sortBy";
     FragmentNavigationBinding binding;
