+++ app/src/main/res/navigation/nav_root.xml
@@ -43,8 +43,8 @@
                 android:name="isTimeSortable"
                 android:defaultValue="false" />
             <argument
-                android:name="coinSortCallback"
-                app:argType="com.hxl.coinfuse.ui.dialogs.coins.CoinSortCallback" />
+                android:name="sortCallback"
+                app:argType="com.hxl.coinfuse.ui.dialogs.SortCallback" />
             <argument
                 android:name="orderBy"
                 app:argType="com.hxl.presentation.OrderBy"
@@ -86,8 +86,8 @@
             app:argType="boolean"
             app:nullable="false"/>
         <argument
-            android:name="coinSortCallback"
-            app:argType="com.hxl.coinfuse.ui.dialogs.coins.CoinSortCallback" />
+            android:name="sortCallback"
+            app:argType="com.hxl.coinfuse.ui.dialogs.SortCallback" />
         <argument
             android:name="orderBy"
             app:argType="com.hxl.presentation.OrderBy"
